Fix warnings
- Unnecessary Casts
- Unnecessary null comparisons

+++ biojava3-structure/src/main/java/org/biojava/bio/structure/align/fatcat/calc/AFPOptimizer.java
@@ -238,7 +238,7 @@ public class AFPOptimizer
          alignScoreUpdate += blockScore[i];
       }
       if(blockNum >= 2)       {
-         alignScoreUpdate += (double)(blockNum - 1) * torsionPenalty;
+         alignScoreUpdate += (blockNum - 1) * torsionPenalty;
       }
 
       afpChain.setBlockGap(blockGap);
